write secrets file in artifact, read same for local development

+++ src/helpers.js
@@ -857,12 +857,24 @@ async function setUpTenantFiles(meshId) {
 	}
 }
 
+/**
+ * This function is to create secrets.yaml in mesh-artifacts for respective meshId. Used for local development run command
+ *
+ * @secretsData secretsData
+ * @meshId meshId
+ */
 async function importSecrets(secretsData, meshId) {
-	const secretsFileName = 'secrets.yaml';
-	const folderPath = path.join(process.cwd(), 'mesh-artifact', meshId);
-	const filePath = path.join(folderPath, secretsFileName);
-
-	fs.writeFileSync(filePath, secretsData);
+	if (!fs.existsSync(path.resolve(process.cwd(), 'mesh-artifact', meshId))) {
+		throw new Error(`Unexpected Error: issue in creating secrets file.`);
+	}
+	try {
+		const secretsFileName = 'secrets.yaml';
+		const folderPath = path.join(process.cwd(), 'mesh-artifact', meshId);
+		const filePath = path.join(folderPath, secretsFileName);
+		fs.writeFileSync(filePath, secretsData);
+	} catch (error) {
+		throw new Error(error.message);
+	}
 }
 
 module.exports = {
